fix: rename app launcher class
