+++ src/main/java/ru/itmo/JavaAdvanced/lesson1/ApplicationRunner.java
@@ -1,18 +1,18 @@
-package ru.itmo.JavaAdvanced;
+package ru.itmo.JavaAdvanced.lesson1;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import ru.itmo.JavaAdvanced.dto.ExampleDto;
+import ru.itmo.JavaAdvanced.lesson1.dto.ExampleDto;
 
 @SpringBootApplication
-public class JavaAdvancedApplication {
+public class ApplicationRunner {
     public static final ExampleDto EXAMPLE_DTO = new ExampleDto();
 
 	public static void main(String[] args) {
 		EXAMPLE_DTO.setId(1L);
         System.out.println(EXAMPLE_DTO.getId());
 
-        SpringApplication.run(JavaAdvancedApplication.class, args);
+        SpringApplication.run(ApplicationRunner.class, args);
 	}
 
 }
