+++ src/test/java/ru/itmo/JavaAdvanced/lesson1/ApplicationRunnerTests.java
@@ -1,10 +1,10 @@
-package ru.itmo.JavaAdvanced;
+package ru.itmo.JavaAdvanced.lesson1;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class JavaAdvancedApplicationTests {
+class ApplicationRunnerTests {
 
 	@Test
 	void contextLoads() {
